server/requirements.txt
flask==3.0.0
flask-socketio==5.3.5
paho-mqtt==1.6.1
plotly==5.18.0

